requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
iniconfig==1.1.1
jsonpickle==1.4.2
kafka-python==2.0.2
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.2
